requirements.txt
Flask==2.2.2
numpy==1.20.3
pandas==1.4.3
Pillow==9.3.0
tensorflow==2.10.0
pybase64==1.2.3
pickle==4.0
enum
typing
